Crate shadow_shim

Re-exports

pub use shimlogger::export as shimlogger_export;

Modules

clone

export

mmap_box

preempt

Native preemption for managed code.

reinit_auxvec_random

shimlogger

signals

syscall

tls

no_std thread-local storage

Structs

ExecutionContextRestorer

Restores an execution context when droped.

Enums

ExecutionContext

Values of this enum describes whether something occurred within the context of the shadow (shim) code, or the application/plugin code.

Constants

EXECUTION_CONTEXT_APPLICATION_CONST

EXECUTION_CONTEXT_SHADOW_CONST

Functions

release_and_exit_current_thread⚠

Release this thread’s shim thread local storage and exit the thread.

simtime