shadow_shim::mmap_box

Struct MmapBox

pub struct MmapBox<T> { /* private fields */ }

Analogous to alloc::boxed::Box, but directly uses mmap instead of a global allocator.

Useful since we donโt currently have a global allocator in the shim, and probably donโt want to install one that makes direct mmap calls for every allocation, since that would be a performance footgun.

We should be able to replace this with alloc::boxed::Box<T> if and when we implement a global allocator suitable for the shim. (Or with alloc::boxed::Box<T, MmapAllocator> when non-global allocators are stabilized)

Implementations

impl<T> MmapBox<T>

pub fn new(x: T) -> Self

pub fn leak(this: MmapBox<T>) -> *mut T