Module shadow_shim::export

Functions

_shim_init_signal_stack

Allocates and installs a signal stack. This is to ensure that our signal handlers have enough stack space; otherwise we can run out in managed processes that use small stacks.

_shim_ipc_wait_for_start_event

_shim_load

_shim_parent_close_stdin

_shim_parent_init_host_shm

_shim_parent_init_ipc⚠

Safety

_shim_parent_init_manager_shm

_shim_preload_only_child_ipc_wait_for_start_event

Wait for start event from shadow, from a newly spawned thread.

shim_api_freeaddrinfo⚠

Safety

shim_api_freeifaddrs⚠

Safety

shim_api_getaddrinfo⚠

Safety

shim_api_getifaddrs⚠

Safety

shim_api_syscall⚠

Safety

shim_freeSignalStack⚠

Safety

shim_hostSharedMem

shim_interpositionEnabled

Whether syscall interposition is currently enabled.

shim_managerSharedMem

shim_processSharedMem

shim_release_and_exit_current_thread⚠

Should be used to exit every thread in the shim.

shim_swapAllowNativeSyscalls

Sets the flag determining whether syscalls are passed through natively, and returns the old value. Typical usage is to set this to the desired value at the beginning of an operation, and restore the old value afterwards.

shim_thisThreadEventIPC⚠

This thread’s IPC channel. Panics if it hasn’t been initialized yet.

shim_threadSharedMem⚠

This thread’s IPC channel. Panics if it hasn’t been initialized yet.